% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treedata}
\alias{treedata}
\title{Density of trees at 20 study plots}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{elevation}{elevation of study plot, factor ("Low", "High")}
  \item{density}{measured density (trees/ha) of trees on the study plot}
}
}
\usage{
treedata
}
\description{
A dataset containing the measured density of trees at low and high elevations
}
\keyword{datasets}
