% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treeHt}
\alias{treeHt}
\title{Height of 32 trees needed for lab 3}
\format{
A data frame with 20 rows and 2 variables:
\describe{
  \item{height}{height, in meters, of each tree}
  \item{species}{species of each tree}
}
}
\usage{
treeHt
}
\description{
A dataset containing the measured height of four species of trees
}
\keyword{datasets}
